[MOV] web,im_livechat: Move clipboard widget to web
