--- addons/im_livechat/static/src/xml/copy_clipboard.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates xml:space="preserve">
-
-    <t t-name="CopyClipboardText">
-        <button class="btn btn-xs btn-primary o_clipboard_button o_btn_text_copy">
-            <span class="fa fa-clipboard">Copy Text</span>
-        </button>
-    </t>
-
-    <t t-name="CopyClipboardChar">
-        <button class="btn btn-xs btn-primary o_clipboard_button o_btn_char_copy">
-            <span class="fa fa-clipboard">Copy Text</span>
-        </button>
-    </t>
-
-</templates>
