+++ addons/web/static/src/scss/fields.scss
@@ -406,4 +406,26 @@
             }
         }
     }
+
+    // Copy to clipboard
+    &.o_field_copy {
+        position: relative;
+        border-radius: 5px;
+        border: 1px solid $btn-primary-border;
+        font-size: $font-size-small;
+        color: $o-brand-optional;
+        font-weight: $badge-font-weight;
+        text-align: center;
+        .o_clipboard_button {
+            &.o_btn_text_copy {
+                position: absolute;
+                top: 0;
+                right: 0;
+            }
+            &.o_btn_char_copy {
+                padding-top: 2px;
+            }
+        }
+    }
+
 }
