+++ addons/web/static/src/xml/base.xml
@@ -930,6 +930,16 @@
         t-att-border="widget.readonly ? 0 : 1"
         t-att-name="widget.name"/>
 </t>
+<t t-name="CopyClipboardText">
+    <button class="btn btn-xs btn-primary o_clipboard_button o_btn_text_copy">
+        <span class="fa fa-clipboard">Copy Text</span>
+    </button>
+</t>
+<t t-name="CopyClipboardChar">
+    <button class="btn btn-xs btn-primary o_clipboard_button o_btn_char_copy">
+        <span class="fa fa-clipboard">Copy Text</span>
+    </button>
+</t>
 <t t-name="FieldBinaryFile">
     <a t-if="widget.mode === 'readonly'" href="javascript:void(0)" class="o_form_uri"/>
